Moves the State class to the socket/model package.
