+++ core/src/main/kotlin/chat/rocket/core/internal/realtime/socket/Login.kt
@@ -2,6 +2,7 @@ package chat.rocket.core.internal.realtime.socket
 
 import chat.rocket.common.model.Token
 import chat.rocket.core.internal.realtime.socket.model.SocketToken
+import chat.rocket.core.internal.realtime.socket.model.State
 import chat.rocket.core.internal.model.TypedResponse
 import chat.rocket.core.internal.realtime.message.loginMethod
 import com.squareup.moshi.Types
