+++ core/src/main/kotlin/chat/rocket/core/internal/realtime/socket/model/State.kt
@@ -1,0 +1,11 @@
+package chat.rocket.core.internal.realtime.socket.model
+
+sealed class State {
+    class Created : State()
+    class Waiting(val seconds: Int) : State()
+    class Connecting : State()
+    class Authenticating : State()
+    class Connected : State()
+    class Disconnecting : State()
+    class Disconnected : State()
+}
